feat: Add crons handler and update database-do to use handler-based routing
